project1/requirements.txt
django
djangorestframework
Pillow
mysqlclient